<chore><move out the widgets to files>
