+++ lib/ui/components/common/default_scaffold.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:modal_progress_hud/modal_progress_hud.dart';
 import 'package:one_piece_platform/ui/components/common/ticket.dart';
-import 'package:one_piece_platform/ui/styles/size_config.dart';
-import 'package:one_piece_platform/ui/text/expandable_text.dart';
+import 'package:one_piece_platform/ui/components/image/thumbnail.dart';
+import 'package:one_piece_platform/ui/screens/user/user_dynamic_header.dart';
 
 import '../../constants.dart';
 
@@ -85,7 +85,7 @@ class DefaultScaffold extends StatelessWidget {
           slivers: <Widget>[
             SliverPersistentHeader(
               floating: true,
-              delegate: MyDynamicHeader(),
+              delegate: UserDynamicHeader(),
             ),
             SliverList(
               delegate: SliverChildListDelegate(
@@ -120,349 +120,4 @@ class DefaultScaffold extends StatelessWidget {
       ),
     );
   }
-}
-
-class HeaderWidget extends StatelessWidget {
-  final String text;
-
-  HeaderWidget(this.text);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: EdgeInsets.all(16.0),
-      child: Text(text),
-      color: Colors.grey[200],
-    );
-  }
-}
-
-class Thumbnail extends StatelessWidget {
-  final String imageURL;
-
-  Thumbnail(this.imageURL);
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(2.0),
-      child: DecoratedBox(
-        decoration: ShapeDecoration(
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.all(Radius.circular(10.0)),
-          ),
-          image: DecorationImage(
-            fit: BoxFit.cover,
-            image: NetworkImage(
-              imageURL,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class MyDynamicHeader extends SliverPersistentHeaderDelegate {
-  int index = 0;
-  final double circleRadius = 100.0;
-  final double circleBorderWidth = 8.0;
-  final double jobTitleWidth = 70.0;
-  final double jobAndAgeFontSize = 17.0;
-  final double nameFontSize = 20.0;
-
-  @override
-  Widget build(
-      BuildContext context, double shrinkOffset, bool overlapsContent) {
-    return LayoutBuilder(builder: (context, constraints) {
-      final Color color = Colors.primaries[index];
-      final double percentage =
-          (constraints.maxHeight - minExtent) / (maxExtent - minExtent);
-      MediaQueryData queryData;
-      queryData = MediaQuery.of(context);
-      if (++index > Colors.primaries.length - 1) index = 0;
-      SizeConfig().init(context);
-
-      return Container(
-        decoration: BoxDecoration(
-          boxShadow: [BoxShadow(blurRadius: 4.0, color: Colors.black45)],
-          gradient: LinearGradient(colors: [Colors.indigo[900], Colors.indigo]),
-        ),
-        height: SizeConfig.safeBlockVertical * 400.0,
-        child: SafeArea(
-          child: Stack(
-            alignment: Alignment.topCenter,
-            children: <Widget>[
-              Padding(
-                padding: EdgeInsets.only(
-                    //clip half of avatar edge
-                    top: circleRadius / 2.0),
-                child: Container(
-                  //replace this Container with your Card description
-                  color: Colors.white,
-                  height: SizeConfig.safeBlockVertical * 400.0,
-
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceAround,
-                    children: <Widget>[
-                      Column(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        crossAxisAlignment: CrossAxisAlignment.center,
-                        mainAxisSize: MainAxisSize.min,
-                        children: [
-                          SizedBox(
-                            height: SizeConfig.safeBlockVertical * 5.0,
-                          ),
-                          Text(
-                            'Evena',
-                            style: TextStyle(
-                              fontSize: nameFontSize,
-                              color: Colors.black,
-                              fontWeight: FontWeight.w500,
-                            ),
-                          ),
-                          Padding(
-                            padding: const EdgeInsets.all(8.0),
-                            child: Container(
-                              child: Padding(
-                                padding: const EdgeInsets.all(3.0),
-                                child: Row(
-                                  children: <Widget>[
-                                    Padding(
-                                      padding: const EdgeInsets.symmetric(
-                                        horizontal: 30.0,
-                                      ),
-                                      child: Column(
-                                        crossAxisAlignment:
-                                            CrossAxisAlignment.center,
-                                        mainAxisAlignment:
-                                            MainAxisAlignment.center,
-                                        mainAxisSize: MainAxisSize.min,
-                                        children: <Widget>[
-                                          Align(
-                                            alignment: Alignment.center,
-                                            child: Container(
-                                              width: jobTitleWidth,
-                                              child: Text(
-                                                'Captain',
-                                                textAlign: TextAlign.center,
-                                                style: TextStyle(
-                                                  fontSize: jobAndAgeFontSize,
-                                                  color: Colors.black,
-                                                  fontWeight: FontWeight.w500,
-                                                ),
-                                              ),
-                                            ),
-                                          ),
-                                        ],
-                                      ),
-                                    ),
-                                    SizedBox(
-                                      height:
-                                          SizeConfig.safeBlockVertical * 2.0,
-                                      child: Container(
-                                        color: Colors.grey[900],
-                                        width: 1,
-                                      ),
-                                    ),
-                                    Padding(
-                                      padding: const EdgeInsets.symmetric(
-                                        horizontal: 30.0,
-                                      ),
-                                      child: Column(
-                                        crossAxisAlignment:
-                                            CrossAxisAlignment.center,
-                                        mainAxisAlignment:
-                                            MainAxisAlignment.center,
-                                        mainAxisSize: MainAxisSize.min,
-                                        children: <Widget>[
-                                          Align(
-                                            alignment: Alignment.center,
-                                            child: Container(
-                                              width: jobTitleWidth,
-                                              child: Text(
-                                                '18',
-                                                textAlign: TextAlign.center,
-                                                style: TextStyle(
-                                                  fontSize: jobAndAgeFontSize,
-                                                  color: Colors.black,
-                                                  fontWeight: FontWeight.w500,
-                                                ),
-                                              ),
-                                            ),
-                                          ),
-                                        ],
-                                      ),
-                                    ),
-                                  ],
-                                ),
-                              ),
-                            ),
-                          ),
-                          Padding(
-                            padding: const EdgeInsets.all(8.0),
-                            child: Container(
-                              decoration: BoxDecoration(
-                                border: Border.all(
-                                  color: Colors.black,
-                                  width: 1,
-                                ),
-                                borderRadius: BorderRadius.circular(12),
-                              ),
-                              child: Padding(
-                                padding: const EdgeInsets.all(3.0),
-                                child: Row(
-                                  children: <Widget>[
-                                    Padding(
-                                      padding: const EdgeInsets.symmetric(
-                                        horizontal: 30.0,
-                                      ),
-                                      child: Column(
-                                        crossAxisAlignment:
-                                            CrossAxisAlignment.center,
-                                        mainAxisAlignment:
-                                            MainAxisAlignment.center,
-                                        mainAxisSize: MainAxisSize.min,
-                                        children: <Widget>[
-                                          Align(
-                                            alignment: Alignment.center,
-                                            child: Text(
-                                              'Followers',
-                                              textAlign: TextAlign.center,
-                                              style: TextStyle(
-                                                fontSize: 14,
-                                                color: Colors.black,
-                                                fontWeight: FontWeight.w500,
-                                              ),
-                                            ),
-                                          ),
-                                          SizedBox(
-                                            height:
-                                                SizeConfig.safeBlockVertical *
-                                                    2.0,
-                                          ),
-                                          Align(
-                                            alignment: Alignment.center,
-                                            child: Text(
-                                              '120',
-                                              textAlign: TextAlign.center,
-                                              style: TextStyle(
-                                                fontSize: 14,
-                                                fontWeight: FontWeight.w400,
-                                              ),
-                                            ),
-                                          ),
-                                        ],
-                                      ),
-                                    ),
-                                    SizedBox(
-                                      height:
-                                          SizeConfig.safeBlockVertical * 10.0,
-                                      child: Container(
-                                        color: Colors.grey[900],
-                                        width: 1,
-                                      ),
-                                    ),
-                                    Padding(
-                                      padding: const EdgeInsets.symmetric(
-                                        horizontal: 30.0,
-                                      ),
-                                      child: Column(
-                                        crossAxisAlignment:
-                                            CrossAxisAlignment.center,
-                                        mainAxisAlignment:
-                                            MainAxisAlignment.center,
-                                        mainAxisSize: MainAxisSize.min,
-                                        children: <Widget>[
-                                          Align(
-                                            alignment: Alignment.center,
-                                            child: Text(
-                                              'Followings',
-                                              textAlign: TextAlign.center,
-                                              style: TextStyle(
-                                                fontSize: 14,
-                                                color: Colors.black,
-                                                fontWeight: FontWeight.w500,
-                                              ),
-                                            ),
-                                          ),
-                                          SizedBox(
-                                            height:
-                                                SizeConfig.safeBlockVertical *
-                                                    2.0,
-                                          ),
-                                          Align(
-                                            alignment: Alignment.center,
-                                            child: Text(
-                                              '152',
-                                              textAlign: TextAlign.center,
-                                              style: TextStyle(
-                                                fontSize: 14,
-                                                fontWeight: FontWeight.w400,
-                                              ),
-                                            ),
-                                          ),
-                                        ],
-                                      ),
-                                    ),
-                                  ],
-                                ),
-                              ),
-                            ),
-                          ),
-                          Container(
-                            width: queryData.size.width,
-                            child: Padding(
-                              padding: const EdgeInsets.all(8.0),
-                              child: Align(
-                                alignment: Alignment.center,
-                                child: ExpandableText(
-                                  "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
-                                ),
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-              Container(
-                width: circleRadius,
-                height: circleRadius,
-                child: Padding(
-                  padding: EdgeInsets.all(circleBorderWidth),
-                  child: DecoratedBox(
-                    decoration: ShapeDecoration(
-                      shape: CircleBorder(),
-                      image: DecorationImage(
-                        fit: BoxFit.cover,
-                        image: NetworkImage(
-                          'https://upload.wikimedia.org/wikipedia/commons/a/a0/Bill_Gates_2018.jpg',
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-              )
-            ],
-          ),
-        ),
-      );
-    });
-  }
-
-  @override
-  bool shouldRebuild(SliverPersistentHeaderDelegate _) => true;
-
-  @override
-  double get maxExtent => SizeConfig.safeBlockVertical != null
-      ? SizeConfig.safeBlockVertical * 70.0
-      : 500.0;
-
-  @override
-  double get minExtent => SizeConfig.safeBlockVertical != null
-      ? SizeConfig.safeBlockVertical * 70.0
-      : 500.0;
 }
